+++ CHANGELOG.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a command to print info about a chip, such as RAM and the number of cores. (#946)
 - ARM:`Session::swo_reader` that returns a wrapping implementation of `std::io::Read` around `Session::read_swo`. (#916)
 
+### Changed
+
+- Chip names are now matched treating an 'x' as a wildcard. (#964)
+
 ### Fixed
 
 - Fixed a panic when cmsisdap probes return more transfers than requested (#922, #923)